Integration between request code and listing model
This is a naive implementation between the request code
and the Listing model. We'll want to clean up the Listing
code and factor out the items that will be common across all
models into a base class, and return nil where appropriate, etc
but this is a nice milestone to reach.