Added notes model and specs
Employed some ruby trickery to make this work. The listing
model has a notes method that builds an anonymous (is that
right?) subclass of the NOtes model in order to properly
set the request path, since those are usually inferred from the
class attribute, but obviously in this case that URI isn't
static.