TASK-866 request_id / request_id_chain logging
Primarily as an aid when debugging, we'll capture the `request_id`
response header that the Spark API servers return. We also add a
`request_id` accessor on both the exception classes, and the response
objects that can be used as follows

```ruby
  begin
    SparkApi.client.get '/account/idontexist'
  rescue SparkApi::NotFound => e
    # e.request_id
  end

  # OR
  SparkApi:003:0> SparkApi.client.get('/my/account').request_id
  => "d349496004aae08763b5c966b2fd1ce0"
```

Additionally, to aid in overall request tracking, if configured, the
`X-Request-Id-Chain` header is passed along. This header is used
internally for logging and instrumentation purposes.

This can be configured in the familiar ways:
```ruby
  SparkApi.configure { |c| c.request_id_chain = 'foo' }

  # OR...
  SparkApi::Client.new(request_id_chain: 'bar')
```