Replace TravisCi with GitHub Actions
This changeset removes the `.travis.yml` configuration and replaces it
with equivalent functionality in `.github/workflows/ci.yml`.

The latter uses GitHub Actions, and beyond running tests for pull
requests and pushes, supports everything needed for us to publish the
gem to Rubygems.org. This keeps the entire workflow on one platform,
rather than what we're doing today where TravisCI is used for continuous
integration, but we're using Jenkins internally to publish the gem.