Name addressable gem as runtime dep
This dependency is already required and has been referenced in code for
some time. Because Webmock requires Addressable and webmock is used in
the test suite, this runtime dependency wasn't surfaced by the
in-project test suite.

This change simply names that gem as a runtime requirement so bundler
can appropriately include it if nothing else happens to.