Name addressable gem as runtime dep
Forward-port of v1.5.7

This dependency is already required and has been referenced in code for
some time. Because Webmock requires Addressable and webmock is used in
the test suite, this runtime dependency wasn't surfaced by the
in-project test suite.

This change simply names that gem as a runtime requirement so bundler
can appropriately include it if nothing else happens to.

Other changes:
 - drop ruby 2.5 and 2.6 from the test suite matrix
 - fix unrelated test breakage in spec
   Not entirely sure what's going on here, but since this fixes the failure
   I'm banking on a deep comparison mismatch when the hash gets recreated.
   Only creating one hash works around it.