BREAKING: remove `compress` option
This change removes the `compress` option from the client entirely.

We use the default faraday adapter, which is `net/http` from stdlib. If
you read [the documentation for that module][net-http-docs], you'll come
across this:

> `Net::HTTP` automatically adds Accept-Encoding for compression of
> response bodies and automatically decompresses gzip and deflate
> responses unless a Range header was sent.

In other words, by default, ruby's net/http module will request gzipped
content and automatically uncompress it if that's what the response
contains.

If instead one sets the `Accept-Encoding` header on a request that's
made with `net/http`, that default behavior is ignored and we're left to
deal with unpacking hte response ourselves as was done in the
`FaradayMiddleware` class.

In other words, whether the `compress` option is used or not, the
behavior has been that we're asking for compressed responses and the
option is effectively ignored. I went back to ruby 1.9.2 and the same
behavior existed there. Given that, I'm confident this option has always
been either ignored or redundant.

[net-http-docs]: https://ruby-doc.org/stdlib-3.0.0/libdoc/net/http/rdoc/Net/HTTP.html#class-Net::HTTP-label-Compression